// src/settings/types.ts
export enum TextBoxIDs {
  SIGNATURE = "signature-textarea",
  AUTOCOMPLETE_COMMENTS = "autocomplete-comments-textarea",
  AUTOCOMPLETE_WORK_NOTES = "autocomplete-work_notes-textarea",
  AUTOCOMPLETE_CLOSE_NOTES = "autocomplete-close_notes-textarea",
}

export enum CheckBoxIDs {
  ALL = "enabled-all",
  //   SERVICENOW = "enabled-service-now",
  SERVICENOW_AUTOCOMPLETE = "enabled-service-now-autocomplete",
  SERVICENOW_MACROS = "enabled-service-now-macros",
  SERVICENOW_MARKDOWN = "enabled-service-now-markdown",
  SERVICENOW_SMART_TEXT = "enabled-service-now-smart-text",
  //   NETDB = "enabled-netdb",
  //   NETDB_MAC_LINKING = "enabled-netdb-mac-linking",
  //   NETDB_HISTORY_HIGHLIGHT = "enabled-netdb-history-highlight",
  //   DHCP_LOG = "enabled-dhcp-log",
}